itoa whith long long

+++ src/libftr/ft_itoa.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-static int	get_size(int n)
+static int	get_size(long long n)
 {
 	int		size;
 
@@ -30,7 +30,7 @@ static int	get_size(int n)
 	return (size);
 }
 
-char		*ft_itoa(int n)
+char		*ft_itoa(long long n)
 {
 	char	*res;
 	int		index;
